requirements.txt
Django==2.1.7
djangorestframework==3.9.2
psycopg2-binary==2.8.2
pytz==2019.1
lxml==4.3.3
